# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Kampungku Housing Website API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AdminLoginBody {
  username: string;
  password: string;
}

export interface AdminLoginResponse {
  success: boolean;
  message: string;
}

export interface AdminSession {
  isAuthenticated: boolean;
  username?: string;
}

export interface ChangePasswordBody {
  currentPassword: string;
  newPassword: string;
}

export interface CctvVerifyBody {
  password: string;
}

export interface CctvVerifyResponse {
  success: boolean;
}

export interface TestResult {
  success: boolean;
  message: string;
}

export interface Feature {
  icon?: string;
  title: string;
  description: string;
}

export interface StatItem {
  value: string;
  label: string;
}

export interface HomepageSettings {
  id: number;
  heroTitle: string;
  heroSubtitle: string;
  heroImageUrl?: string;
  heroBanners?: string[];
  aboutTitle?: string;
  aboutText?: string;
  aboutImageUrl?: string;
  featuresTitle?: string;
  features?: Feature[];
  statsItems?: StatItem[];
  updatedAt?: string;
}

export interface HomepageSettingsBody {
  heroTitle?: string;
  heroSubtitle?: string;
  heroImageUrl?: string;
  heroBanners?: string[];
  aboutTitle?: string;
  aboutText?: string;
  aboutImageUrl?: string;
  featuresTitle?: string;
  features?: Feature[];
  statsItems?: StatItem[];
}

export interface ContactSettings {
  id: number;
  address: string;
  phone?: string;
  email?: string;
  mapEmbedUrl?: string;
  mapLat?: number;
  mapLng?: number;
  officeHours?: string;
  updatedAt?: string;
}

export interface ContactSettingsBody {
  address?: string;
  phone?: string;
  email?: string;
  mapEmbedUrl?: string;
  mapLat?: number;
  mapLng?: number;
  officeHours?: string;
}

export interface EmailSettings {
  id: number;
  gmailUser?: string;
  gmailAppPassword?: string;
  recipientEmail?: string;
  senderName?: string;
  updatedAt?: string;
}

export interface EmailSettingsBody {
  gmailUser?: string;
  gmailAppPassword?: string;
  recipientEmail?: string;
  senderName?: string;
}

export interface DriveSettings {
  id: number;
  appsScriptUrl?: string;
  folderId?: string;
  updatedAt?: string;
}

export interface DriveSettingsBody {
  appsScriptUrl?: string;
  folderId?: string;
}

export interface CctvSettings {
  id: number;
  pagePassword?: string;
  pageTitle?: string;
  pageDescription?: string;
  updatedAt?: string;
}

export interface CctvSettingsBody {
  pagePassword?: string;
  pageTitle?: string;
  pageDescription?: string;
}

export interface DocumentsPageSettings {
  id: number;
  pageTitle?: string;
  pageDescription?: string;
  updatedAt?: string;
}

export interface DocumentsPageSettingsBody {
  pageTitle?: string;
  pageDescription?: string;
}

export type GalleryItemType =
  (typeof GalleryItemType)[keyof typeof GalleryItemType];

export const GalleryItemType = {
  image: "image",
  video: "video",
} as const;

export interface GalleryItem {
  id: number;
  title: string;
  url: string;
  thumbnailUrl?: string;
  type: GalleryItemType;
  description?: string;
  driveFileId?: string;
  createdAt?: string;
}

export interface CreateGalleryItemBody {
  title: string;
  url: string;
  thumbnailUrl?: string;
  description?: string;
  driveFileId?: string;
}

export interface Document {
  id: number;
  title: string;
  description?: string;
  fileUrl: string;
  fileName?: string;
  fileSize?: string;
  fileType?: string;
  driveFileId?: string;
  category?: string;
  createdAt?: string;
}

export interface CreateDocumentBody {
  title: string;
  description?: string;
  fileUrl: string;
  fileName?: string;
  fileSize?: string;
  fileType?: string;
  driveFileId?: string;
  category?: string;
}

export interface ContactMessage {
  id: number;
  name: string;
  email: string;
  phone?: string;
  subject?: string;
  message: string;
  isRead?: boolean;
  createdAt?: string;
}

export interface SendContactMessageBody {
  name: string;
  email: string;
  phone?: string;
  subject?: string;
  message: string;
}

export interface CctvCamera {
  id: number;
  name: string;
  location?: string;
  streamUrl: string;
  isActive?: boolean;
  sortOrder?: number;
  createdAt?: string;
}

export interface CreateCctvCameraBody {
  name: string;
  location?: string;
  streamUrl: string;
  isActive?: boolean;
  sortOrder?: number;
}