# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Kampungku Housing Website API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/admin/login:
    post:
      operationId: adminLogin
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminLoginResponse"
        "401":
          description: Unauthorized
...
  /auth/admin/me:
    get:
      operationId: getAdminMe
      tags: [auth]
      summary: Get current admin session
      responses:
        "200":
          description: Admin session info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"
        "401":
          description: Not authenticated
  /auth/admin/change-password:
    post:
      operationId: adminChangePassword
      tags: [auth]
      summary: Change admin password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordBody"
      responses:
        "200":
          description: Password changed
        "401":
          description: Unauthorized
  /auth/cctv/verify:
    post:
      operationId: verifyCctvPassword
      tags: [auth]
      summary: Verify CCTV page password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CctvVerifyBody"
      responses:
        "200":
          description: Verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CctvVerifyResponse"
        "401":
          description: Wrong password
...
  /gallery/images:
    get:
      operationId: listGalleryImages
      tags: [gallery]
      summary: List gallery images
      responses:
        "200":
          description: List of gallery images
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GalleryItem"
    post:
      operationId: createGalleryImage
      tags: [gallery]
      summary: Add gallery image
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGalleryItemBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GalleryItem"
  /gallery/images/{id}:
    delete:
      operationId: deleteGalleryImage
      tags: [gallery]
      summary: Delete gallery image
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
  /gallery/videos:
    get:
      operationId: listGalleryVideos
      tags: [gallery]
      summary: List gallery videos
      responses:
        "200":
          description: List of gallery videos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GalleryItem"
    post:
      operationId: createGalleryVideo
      tags: [gallery]
      summary: Add gallery video
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGalleryItemBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GalleryItem"
  /gallery/videos/{id}:
    delete:
      operationId: deleteGalleryVideo
      tags: [gallery]
      summary: Delete gallery video
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
  /documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List all documents
      responses:
        "200":
          description: List of documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
    post:
      operationId: createDocument
      tags: [documents]
      summary: Upload document
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDocumentBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
  /documents/{id}:
    delete:
      operationId: deleteDocument
      tags: [documents]
      summary: Delete document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
  /contact/messages:
    get:
      operationId: listContactMessages
      tags: [contact]
      summary: List contact messages
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContactMessage"
    post:
      operationId: sendContactMessage
      tags: [contact]
      summary: Send contact message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendContactMessageBody"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactMessage"
  /contact/messages/{id}:
    delete:
      operationId: deleteContactMessage
      tags: [contact]
      summary: Delete contact message
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
  /cctv/cameras:
    get:
      operationId: listCctvCameras
      tags: [cctv]
      summary: List CCTV cameras
      responses:
        "200":
          description: List of cameras
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CctvCamera"
    post:
      operationId: createCctvCamera
      tags: [cctv]
      summary: Add CCTV camera
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCctvCameraBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CctvCamera"
  /cctv/cameras/{id}:
    put:
      operationId: updateCctvCamera
      tags: [cctv]
      summary: Update CCTV camera
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCctvCameraBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CctvCamera"
    delete:
      operationId: deleteCctvCamera
      tags: [cctv]
      summary: Delete CCTV camera
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
components:
  schemas:
...
    AdminLoginBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password
    AdminLoginResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message
    AdminSession:
      type: object
      properties:
        isAuthenticated:
          type: boolean
        username:
          type: string
      required:
        - isAuthenticated
    ChangePasswordBody:
      type: object
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
      required:
        - currentPassword
        - newPassword
    CctvVerifyBody:
      type: object
      properties:
        password:
          type: string
      required:
        - password
    CctvVerifyResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
...
    GalleryItem:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        url:
          type: string
        thumbnailUrl:
          type: string
        type:
          type: string
          enum: [image, video]
        description:
          type: string
        driveFileId:
          type: string
        createdAt:
          type: string
      required:
        - id
        - title
        - url
        - type
    CreateGalleryItemBody:
      type: object
      properties:
        title:
          type: string
        url:
          type: string
        thumbnailUrl:
          type: string
        description:
          type: string
        driveFileId:
          type: string
      required:
        - title
        - url
    Document:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        fileUrl:
          type: string
        fileName:
          type: string
        fileSize:
          type: string
        fileType:
          type: string
        driveFileId:
          type: string
        category:
          type: string
        createdAt:
          type: string
      required:
        - id
        - title
        - fileUrl
    CreateDocumentBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        fileUrl:
          type: string
        fileName:
          type: string
        fileSize:
          type: string
        fileType:
          type: string
        driveFileId:
          type: string
        category:
          type: string
      required:
        - title
        - fileUrl
    ContactMessage:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        subject:
          type: string
        message:
          type: string
        isRead:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - name
        - email
        - message
    SendContactMessageBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        subject:
          type: string
        message:
          type: string
      required:
        - name
        - email
        - message
    CctvCamera:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        location:
          type: string
        streamUrl:
          type: string
        isActive:
          type: boolean
        sortOrder:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - name
        - streamUrl
    CreateCctvCameraBody:
      type: object
      properties:
        name:
          type: string
        location:
          type: string
        streamUrl:
          type: string
        isActive:
          type: boolean
        sortOrder:
          type: integer
      required:
        - name
        - streamUrl