# lib/api-zod/src/generated/types/cctvCamera.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Kampungku Housing Website API
 * OpenAPI spec version: 0.1.0
 */

export interface CctvCamera {
  id: number;
  name: string;
  location?: string;
  streamUrl: string;
  isActive?: boolean;
  sortOrder?: number;
  createdAt?: string;
}