# lib/api-zod/src/generated/types/createGalleryItemBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Kampungku Housing Website API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateGalleryItemBody {
  title: string;
  url: string;
  thumbnailUrl?: string;
  description?: string;
  driveFileId?: string;
}