# lib/api-zod/src/generated/types/sendContactMessageBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Kampungku Housing Website API
 * OpenAPI spec version: 0.1.0
 */

export interface SendContactMessageBody {
  name: string;
  email: string;
  phone?: string;
  subject?: string;
  message: string;
}