# lib/db/src/schema/documents.ts
import { pgTable, serial, text, timestamp } from "drizzle-orm/pg-core";
import { createInsertSchema } from "drizzle-zod";
import { z } from "zod/v4";

export const documentsTable = pgTable("documents", {
  id: serial("id").primaryKey(),
  title: text("title").notNull(),
  description: text("description").default(""),
  fileUrl: text("file_url").notNull(),
  fileName: text("file_name").default(""),
  fileSize: text("file_size").default(""),
  fileType: text("file_type").default(""),
  driveFileId: text("drive_file_id").default(""),
  category: text("category").default("Umum"),
  createdAt: timestamp("created_at").defaultNow(),
});

export const insertDocumentSchema = createInsertSchema(documentsTable).omit({ id: true, createdAt: true });
export type InsertDocument = z.infer<typeof insertDocumentSchema>;
export type Document = typeof documentsTable.$inferSelect;